# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação - API SuricatoDev (Projeto Excursionistas)",
        "description": "Esta API fornece endpoints para a gestão de caravanas, passageiros e organizadores, realizar reservas, denúncias e avaliações.\n *      🚀 **Principais funcionalidades:**\n *      - Gerenciar caravanas\n *      - Gerenciar Passageiros\n *      - Gerenciar organizadores\n *      - Realização de reservas\n *      - Denúncias e avaliações de caravanas, passageiros e organizadores\n *      🔒 **Segurança:**\n *      - Autenticação segura via token Bearer\n *      - Utiliza Laravel Sanctum para autenticação\n *      - Requer token Bearer para acesso a endpoints protegidos",
        "contact": {
            "name": "Suporte Suricato Dev",
            "email": "filipe.lamego@fatec.sp.gov.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.3.0"
    },
    "paths": {
...
        "/caravanas/{id}/reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Visualizar todas as reservas de uma caravana, somente para o organizador",
                "description": "Este método permite que o organizador visualize todas as reservas de sua caravana.",
                "operationId": "exibirMinhasReservas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da caravana",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservas da caravana encontradas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID da reserva",
                                                "type": "integer"
                                            },
                                            "caravana_id": {
                                                "description": "ID da caravana",
                                                "type": "integer"
                                            },
                                            "passageiro_id": {
                                                "description": "ID do passageiro",
                                                "type": "integer"
                                            },
                                            "status": {
                                                "description": "Status da reserva",
                                                "type": "string",
                                                "enum": [
                                                    "pendente",
                                                    "confirmada",
                                                    "cancelada"
                                                ]
                                            },
                                            "created_at": {
                                                "description": "Data de criação da reserva",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Data da última atualização da reserva",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado"
                    },
                    "404": {
                        "description": "Caravana não encontrada"
                    }
                }
            }
        },
        "/api/caravanas/{id}/reservas": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Reserva de uma caravana",
                "description": "Criação de reserva para o passageiro",
                "operationId": "criarReserva",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da caravana",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "passageiro_id"
                                ],
                                "properties": {
                                    "passageiro_id": {
                                        "description": "ID do passageiro",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva realizada com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Reserva não pode ser realizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Número de vagas excedido!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Autenticação via Bearer Token usando Laravel Sanctum"
            }
        }
    },
...
}